+++ lib/logs/debug_log.dart
@@ -1,9 +1,7 @@
 import 'dart:developer';
 
+import 'package:corextra/logs/enum/log_level.dart';
 import 'package:flutter/foundation.dart';
-
-/// Logging levels
-enum LogLevel { info, warning, error }
 
 /// Logs messages to console in debug mode with optional level.
 ///
@@ -14,7 +12,6 @@ enum LogLevel { info, warning, error }
 /// debugLog('This is an error', level: LogLevel.error);
 /// ```
 void debugLog(String message, {LogLevel level = LogLevel.info}) {
-  if (kDebugMode) {
-    log('[${level.name.toUpperCase()}] $message');
-  }
+  if (!kDebugMode) return;
+  log('[${level.name.toUpperCase()}] $message');
 }